Refactor and add container scanning
